Update QuizService.java
Changed calculateResult method to calculate result in quizService.

Previous method only worked if response list and questions list were in same order. What if response is suffled or list of questions fetched from database is not in order of respose then there answers wont match as we are matching them with index not by id.

Created a HashMap to store key as 'id' and value as 'question' object and store all questions in Map.

Now it became easy to fetch question by "response.getId" and compared both answers.